pyproject.toml
[project]
name = "pyattender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
